Rename Dummy class
