+++ tests/test_ParticleObjectStorer.py
@@ -1,8 +1,8 @@
 import pytest
 import numpy as np
 from sparkx.Particle import Particle
-from sparkx.Loader.DummyLoader import DummyLoader
-from sparkx.Storer.DummyStorer import DummyStorer
+from sparkx.Loader.ParticleObjectLoader import ParticleObjectLoader
+from sparkx.ParticleObjectStorer import ParticleObjectStorer
 
 # Utility function to create particles
 def create_particle(t, x, y, z, mass, E, px, py, pz, pdg, ID, charge, ncoll, form_time, xsecfac, proc_id_origin, proc_type_origin, t_last_coll, pdg_mother1, pdg_mother2, baryon_number, strangeness, weight, status):
@@ -41,24 +41,24 @@ particle_object_list = [[particle1, particle2], [particle1], [particle2]]
 
 # Tests
 
-def test_dummy_storer_initialization():
-    storer = DummyStorer(particle_object_list)
+def test_particleobject_storer_initialization():
+    storer = ParticleObjectStorer(particle_object_list)
     assert storer.particle_list_ == particle_object_list
 
 def test_create_loader():
-    storer = DummyStorer(particle_object_list)
+    storer = ParticleObjectStorer(particle_object_list)
     storer.create_loader(particle_object_list)
-    assert isinstance(storer.loader_, DummyLoader)
+    assert isinstance(storer.loader_, ParticleObjectLoader)
     assert storer.loader_.particle_list_ == particle_object_list
 
 def test_particle_as_list():
-    storer = DummyStorer(particle_object_list)
+    storer = ParticleObjectStorer(particle_object_list)
     particle_list = storer._particle_as_list(particle1)
     expected_list = [1.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 1, 2, 1, 0, 1.1, 1.2, 1, 1, 1.3, 0, 0, 1, 0, 1, 0]
     assert particle_list == expected_list
 
 def test_particle_list():
-    storer = DummyStorer(particle_object_list)
+    storer = ParticleObjectStorer(particle_object_list)
     storer.num_events_ = 3
     storer.num_output_per_event_ = np.array([[0, 2], [1, 1], [2, 1]])
     particle_array = storer.particle_list()
@@ -71,7 +71,7 @@ def test_particle_list():
     assert particle_array == expected_array
 
 def test_particle_list_single_event():
-    storer = DummyStorer([particle_object_list[0]])
+    storer = ParticleObjectStorer([particle_object_list[0]])
     storer.num_events_ = 1
     storer.num_output_per_event_ = np.array([[0, 2]])
     particle_array = storer.particle_list()
@@ -82,7 +82,7 @@ def test_particle_list_single_event():
     assert particle_array == expected_array
 
 def test_print_particle_lists_to_file(tmp_path):
-    storer = DummyStorer(particle_object_list)
+    storer = ParticleObjectStorer(particle_object_list)
     storer.num_events_ = 3
     storer.num_output_per_event_ = np.array([[0, 2], [1, 1], [2, 1]])
